For some reason I do not grasp right now, "c" needs to have uchar type or the method is broken.
git-svn-id: file:///srv/svn/repos/haiku/haiku/trunk@29821 a95241bf-73f2-0310-859d-f6bbb57e9c96

+++ src/apps/terminal/UTF8Char.h
@@ -42,7 +42,7 @@ struct UTF8Char {
 	static int32 ByteCount(char firstChar)
 	{
 		// Note, this does not recognize invalid chars
-		uint32 c = firstChar;
+		uchar c = firstChar;
 		if (c < 0x80)
 			return 1;
 		if (c < 0xe0)
